The apiKey parameter is in initSession
It used to be in the connect method in v2.0 and then in v2.2 we moved it out of connect and into initSession. It turns out, if you had either not included the apiKey in initSession or removed it from connect we would have covered backwards compatibility but the combination of the 2 APIs threw our logic.